requirements.txt
python-telegram-bot[job-queue]==21.10
python-dotenv==1.0.1
google-genai>=1.0.0
aiosqlite>=0.20.0
Pillow>=10.0.0
